Small Refactorings
