+++ lib/in_memory_store.js
