+++ lib/logout.js
@@ -3,7 +3,7 @@ var xpath                 = require('xpath');
 var utils                 = require('./utils');
 var trim_xml              = require('./trim_xml');
 var signers               = require('./signers');
-var Store                 = require('./store');
+var InMemoryStore         = require('./in_memory_store');
 var zlib                  = require('zlib');
 var qs                    = require('querystring');
 
@@ -16,7 +16,7 @@ var REQUEST_EMBEDDED_SIGNATURE_PATH = "//*[local-name(.)='LogoutRequest']/*[loca
 
 module.exports.logout = function (options) {
   options.getSessions = options.getSessions || function(cb){ return cb(null, []); };
-  options.store = options.store || new Store();
+  options.store = options.store || new InMemoryStore();
 
   var prepareAndSendLogoutRequest = function(req, res, next){
     return options.getSessions(function(err, sessions){
@@ -124,50 +124,60 @@ module.exports.logout = function (options) {
   };
 
   return function (req, res, next) {
-    if (req.query.SAMLRequest || req.body.SAMLRequest) {
-      // SP Initated flow - We should store who requested it, so we can reply back
-      return module.exports.parseLogoutRequest(req.query.SAMLRequest || req.body.SAMLRequest, function (err, requestData) {
-        if (err) { return res.send(500, err); }
-        return options.store.store(requestData, function(err){
-          if(err) return next(err);
-                    
-          return prepareAndSendLogoutRequest(req, res, next);          
-        });        
-      });
+    try {
+      // SP Initated flow. Scenario 1 Step 1
+      if (req.query.SAMLRequest || req.body.SAMLRequest) {
+        return parseIncomingLogoutRequest(req.query.SAMLRequest || req.body.SAMLRequest, function (err, requestData) {
+          if (err) { return next(err); }
+          // We should store who requested the logout, so we can reply back with LogoutResponse
+          return options.store.store(requestData, function(err){
+            if (err) { return next(err); }
+
+            return prepareAndSendLogoutRequest(req, res, next);
+          });
+        });
+      }
+
+      // SP Initiated flow in progress, incoming SAMLResponse from SP. Scenario 1 Step 2
+      if (req.query.SAMLResponse || req.body.SAMLResponse) {
+        // If there are sessions left, keep sending LogoutRequest to Session Participants. If not finish
+        validateSamlResponse(req, res, function(err){
+          if (err) {
+            // TODO Mark as partial logout
+          }
+
+          // TODO remove session from incoming LogoutResponse
+          var sessionIndex = ''; //TODO
+          options.store.removeSession(sessionIndex);
+
+          // Continue with next session
+          prepareAndSendLogoutRequest(req, res, next);
+        });
+      }
+
+      // IDP initated - Start flow - In this case we will show a 200 when complete
+      return prepareAndSendLogoutRequest(req, res, next);
+    } catch (e) {
+      return next(e);
     }
-
-    if (req.query.SAMLResponse || req.body.SAMLResponse) {
-      // Logout in progress - If more sessions continue, if not finish      
-      validateSamlResponse(req, res, function(err){
-        if (err) {
-          // Mark as partial logout
-
-        }
-        
-        // remove session
-        var sessionIndex = '';
-        options.store.removeSession(sessionIndex);
-
-        // Continue with next session
-        prepareAndSendLogoutRequest(req, res, next);
-      });
-    }
-
-    // IDP initated - Start flow - In this case we will show a 200 when complete
-    return prepareAndSendLogoutRequest(req, res, next);
   };
 };
 
-module.exports.parseLogoutRequest = function (samlRequest, callback) {
-  utils.parseSamlRequest(samlRequest, function(err, logoutRequestDom) {
-    if (err) return callback(err);
+/**
+ * Parse the SP initiated Logout Request.
+ * This Logout Request is incoming from the SAML SP into the SAML IdP.
+ * @returns {Object} The Logout Request data as a JSON Object
+ */
+function parseIncomingLogoutRequest(samlRequest, callback) {
+  utils.parseSamlRequest(samlRequest, function (err, logoutRequestDom) {
+    if (err) { return callback(err); }
 
     var data = {};
     var issuer = xpath.select("//*[local-name(.)='Issuer' and namespace-uri(.)='urn:oasis:names:tc:SAML:2.0:assertion']/text()", logoutRequestDom);
-    if (issuer && issuer.length > 0) data.issuer = issuer[0].textContent;
+    if (issuer && issuer.length > 0) { data.issuer = issuer[0].textContent; }
 
     var sessionIndex = xpath.select("//*[local-name(.)='SessionIndex']/text()", logoutRequestDom);
-    if (sessionIndex && sessionIndex.length > 0) data.sessionIndex = sessionIndex[0].textContent;
+    if (sessionIndex && sessionIndex.length > 0) { data.sessionIndex = sessionIndex[0].textContent; }
 
     var nameId = xpath.select("//*[local-name(.)='NameID']", logoutRequestDom);
     if (nameId && nameId.length > 0) {
