+++ package.json
@@ -31,8 +31,6 @@
     "express": "~3.1.0",
     "mocha": "~1.8.1",
     "request": "~2.14.0",
-    "xmldom": "~0.1.13",
-    "cheerio": "~0.10.7",
-    "xpath": "0.0.5"
+    "cheerio": "~0.10.7"
   }
 }
